A quick modification to the DB setup files.
I've added a variable that you can change to your preferred email
address.  This will allow you to test the emailing functions on an email
account of your choice.  The customer's emails will all be the email you
specify when you set up the database for testing.  Use the email
variable in the TestCaseDBLauncher.js file.